Extract sphere checkboxes to separate file
