+++ Makefile
@@ -6,7 +6,7 @@
 #    By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2020/02/29 13:58:56 by oadhesiv          #+#    #+#              #
-#    Updated: 2021/09/11 20:26:55 by oadhesiv         ###   ########.fr        #
+#    Updated: 2021/09/11 21:22:18 by oadhesiv         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -39,7 +39,7 @@ FILES_LOOP :=	loop_hook_flow.c loop_hook_invalidate.c \
 FILES_LOOP := $(addprefix loop/, $(FILES_LOOP))
 
 FILES_IMGUI :=	controls.c imgui_init.c windows.c window_controls.c \
-				window_controls_spheres.c
+				window_controls_spheres.c window_controls_spheres_checkbox.c
 FILES_IMGUI := $(addprefix imgui/, $(FILES_IMGUI))
 
 FILES_SDL :=	init_sdl.c sdl_events.c sdl_gl_attributes.c sdl_clear_texture.c
