+++ includes/imgui/window_controls_spheres_checkbox.h
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   window_controls_spheres_checkbox.h                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/09/11 21:20:43 by oadhesiv          #+#    #+#             */
+/*   Updated: 2021/09/11 21:21:31 by oadhesiv         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef WINDOW_CONTROLS_SPHERES_CHECKBOX_H
+# define WINDOW_CONTROLS_SPHERES_CHECKBOX_H
+
+# include "rtv.h"
+
+void	imgui_sphere_local_checkboxes(t_rtv *rtv, t_byte idx);
+
+#endif
