+++ srcs/imgui/window_controls_spheres.c
@@ -6,7 +6,7 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/07 20:59:42 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/11 21:09:36 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/09/11 21:21:56 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 #include "rtv.h"
 #include "imgui/controls.h"
 #include "loop/loop_hook_invalidate.h"
+#include "imgui/window_controls_spheres_checkbox.h"
 
 const char		*g_sphere_labels[10] = {
 	"Sphere #0",
@@ -31,8 +32,6 @@ const char		*g_sphere_labels[10] = {
 
 float			g_color[MAX_SPHERES][3];
 bool			g_enabled[MAX_SPHERES];
-const t_byte	g_tex_min = 0;
-const t_byte	g_tex_max = MAX_TEXTURES - 1;
 
 static void	imgui_init_sphere_globals(t_rtv *rtv)
 {
@@ -61,57 +60,6 @@ static void	imgui_sphere_enabled(t_rtv *rtv, t_byte idx)
 	else
 		rtv->spheres[idx].traits -= TRAIT_EXISTS;
 	rtv->flags |= FLAG_REDRAW;
-}
-
-static void	imgui_sphere_normale_disruption(t_rtv *rtv, t_byte idx)
-{
-	static bool	normal_sin[MAX_SPHERES];
-
-	if (!(igCheckbox("normale disruption", normal_sin + idx)))
-		return ;
-	if (normal_sin[idx])
-		rtv->spheres[idx].traits |= TRAIT_NORMAL_SIN;
-	else
-		rtv->spheres[idx].traits -= TRAIT_NORMAL_SIN;
-	rtv->flags |= FLAG_REDRAW;
-}
-
-static void	imgui_sphere_textured(t_rtv *rtv, t_byte idx)
-{
-	static bool	textured[MAX_SPHERES];
-
-	if (igCheckbox("textured", textured + idx))
-	{
-		if (textured[idx])
-			rtv->spheres[idx].traits |= TRAIT_TEXTURED;
-		else
-			rtv->spheres[idx].traits -= TRAIT_TEXTURED;
-		rtv->flags |= FLAG_REDRAW;
-	}
-	if (igSliderScalar("texture_id", ImGuiDataType_U8, \
-		&rtv->spheres[idx].texture_id, &g_tex_min, &g_tex_max, "%d", 0) \
-	)
-		rtv->flags |= FLAG_REDRAW;
-}
-
-static void	imgui_sphere_color_checkerboard(t_rtv *rtv, t_byte idx)
-{
-	static bool	checkerboard[MAX_SPHERES];
-
-	if (!(igCheckbox("color disruption: checkboard", checkerboard + idx)))
-		return ;
-	if (checkerboard[idx])
-		rtv->spheres[idx].traits |= TRAIT_CHECKERBOARD;
-	else
-		rtv->spheres[idx].traits -= TRAIT_CHECKERBOARD;
-	rtv->flags |= FLAG_REDRAW;
-}
-
-static void	imgui_sphere_local_checkboxes(t_rtv *rtv, t_byte idx)
-{
-	imgui_sphere_normale_disruption(rtv, idx);
-	imgui_sphere_textured(rtv, idx);
-	imgui_sphere_color_checkerboard(rtv, idx);
 }
 
 static void	imgui_sphere_group(t_rtv *rtv, t_byte idx)
